requirements.txt
aiohttp
asyncio
discord.py
google
google-api-python-client
google-auth-oauthlib
gspread
python-dotenv

